Break up the logic
